requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pyjwt==2.7.0
